FUCK JACK v2
 - Please see updated "README.md"
 - Borrowed code from my other project "EZ AOSP"
 - A much friendlier interface
 - Formatting fixes
 - Whiptail dialog boxes
 - Add colors
 - Kill some white space
 - Remove build/envsetup.sh, lunch, and make commands (this is a tool to be used BEFORE doing all of this)

TO DO:
 - Automatically read user's RAM amount and set arguments based on that
 - Create "simple" script to be placed into user build scripts

+++ build.sh
@@ -4,7 +4,9 @@
 # The "Fuck Jack Build Script"
 #
 # Written by Michael S Corigliano (Mike Criggs) (michael.s.corigliano@gmail.com)
+# Taken & adapted from my EZ AOSP tool, available here: https://github.com/mikecriggs/ez-aosp
 #
+# Usage: ./fuckjack.sh
 #
 # This software is licensed under the terms of the GNU General Public
 # License version 2, as published by the Free Software Foundation, and
@@ -15,37 +17,57 @@
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU General Public License for more details.
 #
-# Please maintain this if you use this script or any part of it
+
+# Colors for terminal output
+  BLDRED="\033[1m""\033[31m"
+  RST="\033[0m"
+
+# Do we want to use NINJA?
+  if(whiptail --title "FUCK JACK" --yesno "Would you like to use NINJA to compile the $ROM source code?" 10 70) then
+      echo -e ${BLDRED}"Compiling $ROM with NINJA..."${RST}
+      sleep 3
+  else
+      echo -e ${BLDRED}"Not compiling $ROM with NINJA..."${RST}
+      sleep 3
+      export USE_NINJA=false
+  fi
+
+# Do we need to use JACK workarounds?
+  if(whiptail --title "FUCK JACK" --yesno "If your machine has less than 16GB of RAM, we may need to use some JACK workarounds. Would you like to use these?" 30 70) then
+      echo -e ${BLDRED}"Using JACK workarounds..."${RST}
+      sleep 3
+      ./prebuilts/sdk/tools/jack-admin kill-server
+      rm -rf ~/.jack*
+
+# IMPORTANT! IMPORTANT! IMPORTANT! IMPORTANT! IMPORTANT! IMPORTANT! IMPORTANT! IMPORTANT! IMPORTANT! IMPORTANT!
 #
+# Replace the value in the first of "ANDROID_JACK_VM_ARGS" with half of your ram
+# (i.e.: if you have 8GB of ram, use "-Xmx4g")
 #
-# The purpose of this script is to work around JACK and NINJA, which have been
-# broken in AOSP as of android-7.0.
+      export ANDROID_JACK_VM_ARGS="-Xmx4g -XX:+TieredCompilation -Dfile.encoding=UTF-8"
 #
-# Usage: ./build.sh <DEVICE>
-#
+# IMPORTANT! IMPORTANT! IMPORTANT! IMPORTANT! IMPORTANT! IMPORTANT! IMPORTANT! IMPORTANT! IMPORTANT! IMPORTANT!
 
-DEVICE="$1"
+      export SERVER_NB_COMPILE=2
+      export ANDROID_JACK_VM_ARGS=$JACK_SERVER_VM_ARGUMENT
 
-# Tell the environment not to use NINJA
-	export USE_NINJA=false
+# You may need to uncomment the following command if Java tells you to use it.
+#     ulimit -n unlimited
 
-# Delete the JACK server located in /home/<USER>/.jack*
-	rm -rf ~/.jack*
+      ./prebuilts/sdk/tools/jack-admin install-server prebuilts/sdk/tools/jack-launcher.jar prebuilts/sdk/tools/jack-server-4.8.ALPHA.jar
+      ./prebuilts/sdk/tools/jack-admin start-server
+  else
+      echo -e ${BLDRED}"Not using JACK workarounds..."${RST}
+      sleep 3
+  fi
 
-# Resize the JACK Heap size
-	export ANDROID_JACK_VM_ARGS="-Xmx4g -Dfile.encoding=UTF-8 -XX:+TieredCompilation"
-
-# Restart the JACK server
-	./prebuilts/sdk/tools/jack-admin kill-server
-	./prebuilts/sdk/tools/jack-admin start-server
-
-# Optionally, you may want to clear CCACHE if you still have issues
-#	ccache -C
-
-# Make a clean build, building dirty after you have had jack issues may result in a failed build
-	make clean
-
-# Compile the build
-	. build/envsetup.sh
-	lunch du_$DEVICE-userdebug
-	make bacon
+# Do we want to build clean?
+  if(whiptail --title "FUCK JACK" --yesno "Would you like to build clean? If you used JACK workarounds, it is recommended to do so or your build may fail." 30 70) then
+      echo -e ${BLDRED}"Building clean..."${RST}
+      sleep 3
+      make clean
+      clear
+  else
+      echo -e ${BLDRED}"Not building clean..."${RST}
+      sleep 3
+  fi
